\name{PatchProcess}
\alias{PatchProcess}
\alias{PatchProcess}
\docType{package}
\title{
Application with a graphical user interface to perform 
Fragmention algorithm and Syntax analysis on digital images.
}
\description{
The package provides a set of tools to perform unsupervised classification on images, namely
fragmentation and syntax algorithm.
}
\details{
  The overall application takes jpg images as inputs converted on a greyscale as input.
}
\author{
Olivier Catoni, Gautier Appert

Maintainer: Olivier Catoni, Gautier Appert
}
\references{
  InformationInformation k-means, fragmentation and syntax analysis:
A new approach to unsupervised Machine Learning.
Phd thesis 2020.
}
\keyword{Fragmentation, Syntax, Patches, Labels }
\seealso{

}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
